Refactors image slider to be generic component. Adds homepageHelper. Passes dict of images to slider comp.
