+++ src/components/imageSlider/ImageSlider.js
@@ -2,28 +2,15 @@ import React, { useState, useEffect } from "react";
 import "./imageSlider.scss";
 
 /**
+ * Creates image slider based on a dict of images with each entry of the form:
+ * {id: (int), url: (string), alt: (string),}
+ * @param props Dictionary of images
  * @returns {JSX.Element}
  * @constructor
  */
-export default function ImageSlider() {
+export default function ImageSlider(props) {
+    const images = props.images;
     const [currentImageID, setCurrentImageID] = useState(1);
-    const images = [
-        {
-            id: 1,
-            url: "https://res.cloudinary.com/djv69vvs7/image/upload/c_scale,w_771/v1628183232/samples/greg-rakozy-oMpAz-DN-9I-unsplash_fsnc7t.jpg",
-            alt: "placeholder",
-        },
-        {
-            id: 2,
-            url: "https://res.cloudinary.com/djv69vvs7/image/upload/c_scale,w_817/v1628183234/samples/john-fowler-7Ym9rpYtSdA-unsplash_ebolox.jpg",
-            alt: "placeholder 2",
-        },
-        {
-            id: 3,
-            url: "https://res.cloudinary.com/djv69vvs7/image/upload/v1628183342/samples/guillermo-ferla-Oze6U2m1oYU-unsplash_mjn1oa.jpg",
-            alt: "placeholder 3",
-        },
-    ];
 
     useEffect(() => {
         const interval = setInterval(() => {
@@ -52,14 +39,10 @@ export default function ImageSlider() {
             <div className="slider-dots">
                 {images.map((button) => (
                     <button
-                        style={
-                            currentImageID === button.id
-                                ? { background: "white" }
-                                : {}
-                        }
+                        style={currentImageID === button.id ? { background: "white" } : {}}
                         key={button.id}
                         onClick={() => sliderHandler(button.id)}
-                    ></button>
+                    />
                 ))}
             </div>
         </div>
