+++ src/components/imageSlider/imageSlider.scss
