+++ src/pages/home/homepageHelper.js
@@ -1,0 +1,21 @@
+/**
+ * Images for home page introduction slider
+ * @type {[{alt: string, id: number, url: string}, {alt: string, id: number, url: string}, {alt: string, id: number, url: string}]}
+ */
+export const homeSliderImages = [
+	{
+		id: 1,
+		url: "https://res.cloudinary.com/djv69vvs7/image/upload/c_scale,w_771/v1628183232/samples/greg-rakozy-oMpAz-DN-9I-unsplash_fsnc7t.jpg",
+		alt: "placeholder",
+	},
+	{
+		id: 2,
+		url: "https://res.cloudinary.com/djv69vvs7/image/upload/c_scale,w_817/v1628183234/samples/john-fowler-7Ym9rpYtSdA-unsplash_ebolox.jpg",
+		alt: "placeholder 2",
+	},
+	{
+		id: 3,
+		url: "https://res.cloudinary.com/djv69vvs7/image/upload/v1628183342/samples/guillermo-ferla-Oze6U2m1oYU-unsplash_mjn1oa.jpg",
+		alt: "placeholder 3",
+	},
+];
